Move promoteObject() to GCImpl from HeapImpl.
