+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/GreyToBlackObjRefVisitor.java
@@ -116,7 +116,7 @@ final class GreyToBlackObjRefVisitor implements ObjectReferenceVisitor {
             // Promote the Object if necessary, making it at least grey, and ...
             Object obj = p.toObject();
             assert innerOffset < LayoutEncoding.getSizeFromObject(obj).rawValue();
-            Object copy = HeapImpl.getHeapImpl().promoteObject(obj, header);
+            Object copy = GCImpl.getGCImpl().promoteObject(obj, header);
             if (copy != obj) {
                 // ... update the reference to point to the copy, making the reference black.
                 counters.noteCopiedReferent();
